Added ability for user to leave team
Based on clicking button with id='leaveTeam'.  Button is disabled unless a player has clicked to join a team -- this will need updating in future, as it does not currently know if a user is in a game without clicking to join.  Hence, if a user is still logged in, leaves the lobby and re-enters, they'll still be in a team but the leave button will be disabled.  Same applies to switching servers with the select dropdown.